Move test to its own project.
