+++ test/Microsoft.ML.Tests/Microsoft.ML.Tests.csproj
@@ -51,6 +51,5 @@
     <PackageReference Include="Microsoft.ML.TensorFlow.TestModels" Version="$(MicrosoftMLTensorFlowTestModelsVersion)" />
     <PackageReference Include="Microsoft.ML.Onnx.TestModels" Version="$(MicrosoftMLOnnxTestModelsVersion)" />
     <PackageReference Include="Microsoft.ML.TestModels" Version="$(MicrosoftMLTestModelsPackageVersion)" />
-    <PackageReference Include="SharpZipLib" Version="1.2.0" />
   </ItemGroup>
 </Project>
